Introduce `QueryColumnsProvider`
